[YS-151] refactor: 인증 코드 입력 컴포넌트 분리
