+++ src/app/join/components/UnivAuthInput/UnivAuthInput.tsx
@@ -1,14 +1,10 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
 import {
-  authCodeButton,
-  authInputContainer,
-  authTimerWrapper,
   editButton,
   errorMessage,
   inputContainer,
   required,
-  sendAgainButton,
   univAuthButton,
   univInputWrapper,
 } from './UnivAuthInput.styles';
@@ -16,7 +12,7 @@ import useSendUnivAuthCodeMutation from '../../hooks/useSendUnivAuthCodeMutation
 
 import EmailToast from '../EmailToast/EmailToast';
 import { FormInput } from '../../JoinPage.types';
-import { formatAuthTimer } from '../../JoinPage.utils';
+import AuthCodeInput from './AuthCodeInput/AuthCodeInput';
 
 const TEN_MINUTE_SEC = 600;
 const ONE_SEC = 1000;
@@ -134,47 +130,10 @@ const UnivAuthInput = ({ isUnivVerify, handleVerifyUniv }: UnivAuthInputProps) =
       {errors.univEmail && <span css={errorMessage}>{errors.univEmail.message}</span>}
       {sendError && <span css={errorMessage}>{sendError.message}</span>}
       {isEmailSent && (
-        <Controller
-          name="authCode"
-          control={control}
-          render={({ field }) => {
-            const isDisabled =
-              Boolean(errors.authCode) || !watch('authCode') || watch('authCode').length < 6;
-
-            return (
-              <div css={authInputContainer}>
-                <div css={univInputWrapper}>
-                  <input
-                    {...field}
-                    placeholder="인증번호 6자리 입력"
-                    type="number"
-                    disabled={isUnivVerify}
-                    onChange={(e) => {
-                      if (e.target.value.length <= 6) {
-                        field.onChange(e);
-                      }
-                    }}
-                  />
-                  {!isUnivVerify && (
-                    <div css={authTimerWrapper}>
-                      <span>{formatAuthTimer(authTimer)}</span>
-                      <button
-                        type="button"
-                        css={authCodeButton}
-                        disabled={isDisabled}
-                        onClick={handleVerifyUniv}
-                      >
-                        인증
-                      </button>
-                    </div>
-                  )}
-                </div>
-                <button type="button" css={sendAgainButton}>
-                  인증번호 재전송
-                </button>
-              </div>
-            );
-          }}
+        <AuthCodeInput
+          isUnivVerify={isUnivVerify}
+          handleVerifyUniv={handleVerifyUniv}
+          authTimer={authTimer}
         />
       )}
       <EmailToast
